move all SMR impls to smr.rs
